feat: allow base package name and version as inputs
this adds git ref as a mechanism to assign version to a
Cartfile.resolved within a directory. The name of the package
will be the dirname where the `Cartfile.resolved` is located.

This avoids the scenario where the monitored git output will
show `_root@0.0.0` and provides slightly better user experience
with the cost of having git available locally.